requirements.txt
langchain
python-dotenv
requests 